Group No. 13 ShopSmartAI/simple_requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-MySQLdb==1.0.1
bcrypt==4.0.1
mysqlclient==2.2.0
pandas==2.0.3
scikit-learn==1.3.0
numpy==1.24.3
